(PCP-424) Connection must not return before connection_timeout
Previous to this commit, the connectAndWait method was returning
immediately in case the WS connection timeout was set to a value lesser
than 1 s; here we fix this.

Adding unit tests to check that the dtor does not hang in case the
configured WebSocket connection timeout is ~ 1 ms.

+++ lib/src/connector/connection.cc
@@ -274,8 +274,9 @@ void Connection::close(CloseCode code, const std::string& reason) {
 //
 
 void Connection::connectAndWait() {
+    // TODO(ale): use Timer::elapsed_milliseconds once it's released
     static auto connection_timeout_s =
-        static_cast<int>(client_metadata_.connection_timeout / 1000);
+        static_cast<double>(client_metadata_.connection_timeout) / 1000.0;
     connect_();
     lth_util::Timer timer {};
     while (connection_state_.load() != ConnectionState::open
